Move Limit to types and remove old files
